update benchmark data (#10)

+++ scripts/kill_servers.sh
@@ -2,4 +2,8 @@
 
 ps -ef | grep reciever | grep -v grep
 pid=$(ps -ef | grep reciever | grep -v grep | awk '{print $2}')
-kill -9 $pid
+
+# kill if server exist
+if [[ -n "$pid" ]]; then
+  kill -9 $pid
+fi
